Use better mocks and renamed svg folder
