+++ src/components/dashboard/suggested.tsx
@@ -1,19 +1,18 @@
 import { Link } from "expo-router";
 import { Pressable, StyleSheet, Text, View } from "react-native";
 
-import BottomGet from "@/assets/svgs/bottom-get";
-import Placeholder from "@/assets/svgs/placeholder";
+import BottomGet from "@/assets/svg/bottom-get";
+import Placeholder from "@/assets/svg/placeholder";
 import Picture from "@/components/picture";
+import { suggested } from "@/mocks/fixtures";
 import { spectrum } from "@/theme";
-
-import { data10 as data } from "./data";
 
 const width = 72;
 
 export default function Favorites() {
   return (
     <View style={styles.container}>
-      {data.map(({ uuid, name, thumbnail, has_campaign }) => (
+      {suggested.map(({ uuid, name, thumbnail, has_campaign }) => (
         <Link href="/" asChild key={uuid}>
           <Pressable>
             <View style={styles.favoriteContainer}>
