+++ src/utils/chunk.ts
